Add migration run seederPrivilages * fix bug edit shipping info. * add reset pre-order

+++ modules/product/src/Http/Controllers/ProductAttributeController.php
@@ -2,11 +2,13 @@
 
 namespace Modules\Product\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use Modules\Product\ProductVariantAttribute;
 use DataTables;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Modules\Product\ProductVariantAttribute;
 use Validator;
+use Auth;
 
 class ProductAttributeController extends Controller
 {
@@ -14,16 +16,16 @@ class ProductAttributeController extends Controller
     {
         return view('product::product-variant.index');
     }
-    
+
     public function create()
     {
         $data = [
-            'title'=>'Add New Product Variant',
-            'back'=> route('product-variant.index'),
+            'title' => 'Add New Product Variant',
+            'back' => route('product-variant.index'),
         ];
-        return view('product::product-variant.create',compact('data'));
+        return view('product::product-variant.create', compact('data'));
     }
-    
+
     public function store(Request $request)
     {
         $request->validate([
@@ -38,35 +40,35 @@ class ProductAttributeController extends Controller
     {
         $productVariant = ProductVariantAttribute::findOrFail($id);
         $data = [
-            'title'=>'Details Product Variant',
-            'back'=> route('product-variant.index'),
+            'title' => 'Details Product Variant',
+            'back' => route('product-variant.index'),
         ];
-        return view('product::product-variant.show',compact('productVariant','data'));
+        return view('product::product-variant.show', compact('productVariant', 'data'));
     }
 
     public function edit(int $id)
     {
         $productVariant = ProductVariantAttribute::findOrFail($id);
         $data = [
-            'title'=>'Edit Product Variant',
-            'back'=> route('product-variant.index'),
+            'title' => 'Edit Product Variant',
+            'back' => route('product-variant.index'),
         ];
         $categories = [];
         if (class_exists('\Modules\ProductCategory\ProductCategory')) {
             $categories = \Modules\ProductCategory\ProductCategory::whereNull('parent_id')
-                            ->with('subcategories')
-                            ->get();            
+                ->with('subcategories')
+                ->get();
         }
-        return view('product::product-variant.edit',compact('productVariant','data','categories'));
+        return view('product::product-variant.edit', compact('productVariant', 'data', 'categories'));
     }
 
-    public function update(Request $request,int $id)
+    public function update(Request $request, int $id)
     {
         $request->validate([
-            'attribute' => 'required'
+            'attribute' => 'required',
         ]);
         $productVariantAttr = ProductVariantAttribute::findOrFail($id);
-        $productVariantAttr->update($request->only('attribute','value'));
+        $productVariantAttr->update($request->only('attribute', 'value'));
 
         return redirect()->route('product-variant.index');
     }
@@ -75,24 +77,33 @@ class ProductAttributeController extends Controller
     {
         $productVariant = ProductVariantAttribute::findOrFail($id);
         if ($productVariant->delete()) {
-            return response()->json(['data'=>1]);
-        }else{
-            return response()->json(['data'=>0]);
+            return response()->json(['data' => 1]);
+        } else {
+            return response()->json(['data' => 0]);
         }
     }
 
     public function getAllDatatables()
     {
-        $variant = ProductVariantAttribute::select(['id','attribute','value'])->get();
+        $variant = ProductVariantAttribute::select(['id', 'attribute', 'value'])->get();
         return Datatables::of($variant)
-        ->addColumn('action', function ($data) {
-            $button = '<a href="'.route('product-variant.edit',$data->id).'" class="btn btn-table circle-table edit-table"';
-            $button .='data-toggle="tooltip" data-placement="top" title="Edit"></a>';
-            $button .='<a href="#" onclick="deleteItem('.$data->id.')" class="btn btn-table circle-table delete-table" data-toggle="tooltip"';
-            $button .=' data-placement="top" title="Delete"></a>';
-            return $button;
-        })
-        ->make(true);
+            ->addColumn('action', function ($data) {
+                
+                $user = Auth::user();
+                $button = '';
+
+                if (Gate::forUser($user)->allows('edit-variant')) {
+                    $button .= '<a href="' . route('product-variant.edit', $data->id) . '" class="btn btn-table circle-table edit-table"';
+                    $button .= 'data-toggle="tooltip" data-placement="top" title="Edit"></a>';
+                }
+
+                if (Gate::forUser($user)->allows('delete-variant')) {
+                    $button .= '<a href="#" onclick="deleteItem(' . $data->id . ')" class="btn btn-table circle-table delete-table" data-toggle="tooltip"';
+                    $button .= ' data-placement="top" title="Delete"></a>';
+                }
+                return $button;
+            })
+            ->make(true);
     }
 
     public function ajaxAddVariant(Request $request)
@@ -105,16 +116,16 @@ class ProductAttributeController extends Controller
         if ($validator->fails()) {
             return response()->json($validator->messages());
         }
-        
+
         $variant = ProductVariantAttribute::create($request->except('_token'));
-        
+
         return response()->json($variant);
     }
 
     public function ajaxGetAllVariant()
     {
         $variant = ProductVariantAttribute::all();
-        
+
         return response()->json($variant);
     }
 
@@ -127,16 +138,16 @@ class ProductAttributeController extends Controller
         if ($validator->fails()) {
             return response()->json($validator->messages());
         }
-        
-        $variant    = ProductVariantAttribute::where('attribute',$request->attribute)->get()->first();
+
+        $variant = ProductVariantAttribute::where('attribute', $request->attribute)->get()->first();
         $attributes = [];
         if (!is_null($variant)) {
-            $attributes = explode(',',$variant->value);
+            $attributes = explode(',', $variant->value);
         }
         return response()->json($attributes);
     }
 
-    public function ajaxAddByIDVariant(Request $request,int $id)
+    public function ajaxAddByIDVariant(Request $request, int $id)
     {
         $validator = Validator::make($request->all(), [
             'value' => 'required',
@@ -146,7 +157,7 @@ class ProductAttributeController extends Controller
             return response()->json($validator->messages());
         }
         $variant    = ProductVariantAttribute::find($id);
-        $new_value  = $variant->value.','.$request->value;
+        $new_value  = $variant->value . ',' . $request->value;
         $variant->value = $new_value;
         $variant->update();
         return response()->json($variant);
